# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StreamPay API",
    "version": "2.0.0",
    "description": "REST API for the StreamPay payment-streaming platform on Stellar.\n\nAll error responses use the canonical envelope:\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"Human-readable detail.\", \"request_id\": \"req_...\" } }\n```\n\nv1 stream paths (`/api/streams`) are deprecated and will be removed. Use `/api/v2/streams`."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Obtain a token via POST /api/auth/wallet, then pass it as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code.",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable description safe to display to clients.",
                "example": "The requested stream does not exist."
              },
              "request_id": {
                "type": "string",
                "description": "Opaque identifier for this request, useful for log correlation.",
                "example": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "StreamV1": {
        "type": "object",
        "description": "v1 payment stream shape. **Deprecated** — use `StreamV2` via `/api/v2/streams`.",
        "required": ["id", "recipient", "rate", "schedule", "status", "nextAction", "token", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique stream identifier.",
            "example": "stream-abc12345"
          },
          "recipient": {
            "type": "string",
            "description": "Stellar public key of the payment recipient.",
            "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
          },
          "rate": {
            "type": "string",
            "description": "Numeric payment amount per schedule interval.",
            "example": "120"
          },
          "schedule": {
            "type": "string",
            "enum": ["second", "minute", "hour", "day", "week", "month", "year"],
            "description": "Payment interval.",
            "example": "month"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "ended"],
            "description": "Current lifecycle state of the stream."
          },
          "nextAction": {
            "type": ["string", "null"],
            "description": "The next permitted action for the authenticated user.",
            "example": "start"
          },
          "token": {
            "type": "string",
            "description": "Stellar asset code.",
            "example": "XLM"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp.",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 last-updated timestamp.",
            "example": "2024-01-15T10:00:00.000Z"
          }
        }
      },
      "StreamV2": {
        "type": "object",
        "required": ["id", "recipient", "rate", "status", "allowed_actions", "created_at", "settlement"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique stream identifier.",
            "example": "stream_lf3k9z"
          },
          "recipient": {
            "type": "string",
            "description": "Stellar public key of the payment recipient.",
            "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
          },
          "rate": {
            "type": "string",
            "description": "Human-readable payment rate.",
            "example": "120 XLM/month"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "ended"],
            "description": "Current lifecycle state of the stream."
          },
          "allowed_actions": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Actions the authenticated user may perform on this stream. Replaces v1 `actions`.",
            "example": ["pause", "stop"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp. Replaces v1 `createdAt`.",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "settlement": {
            "oneOf": [
              {
                "type": "object",
                "required": ["settled_at", "amount", "currency"],
                "properties": {
                  "settled_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-02-01T12:00:00.000Z"
                  },
                  "amount": {
                    "type": "string",
                    "example": "360 XLM"
                  },
                  "currency": {
                    "type": "string",
                    "example": "XLM"
                  }
                }
              },
              { "type": "null" }
            ],
            "description": "Settlement details once the stream has been settled, otherwise `null`. Clients must handle `null` explicitly."
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request — invalid input.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request body must include 'recipient' and 'rate'.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized — missing or invalid bearer token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Bearer token required.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "422": {
        "description": "Unprocessable entity — request body failed schema validation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "One or more fields are invalid.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "An unexpected error occurred.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/streams": {
      "get": {
        "operationId": "listStreams",
        "summary": "List payment streams (v1, deprecated)",
        "description": "Returns a cursor-paginated list of streams owned by the authenticated identity. **Deprecated** — use `GET /api/v2/streams` instead.",
        "deprecated": true,
        "tags": ["Streams v1"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of streams to return (1–100, default 20).",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20, "example": 20 }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque pagination cursor from a previous response `meta.nextCursor` field.",
            "schema": { "type": "string", "example": "c3RyZWFtLWFiYzEyMzQ1" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter results to streams with this status.",
            "schema": { "type": "string", "enum": ["draft", "active", "paused", "ended"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated stream list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "links", "meta"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/StreamV1" }
                    },
                    "links": {
                      "type": "object",
                      "required": ["self"],
                      "properties": {
                        "self": { "type": "string", "example": "/api/v1/streams?limit=20" }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": ["hasNext", "nextCursor", "total"],
                      "properties": {
                        "hasNext": { "type": "boolean", "example": false },
                        "nextCursor": { "type": ["string", "null"], "example": null },
                        "total": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                },
                "examples": {
                  "single-active-stream": {
                    "summary": "One active stream, no further pages",
                    "value": {
                      "data": [
                        {
                          "id": "stream-abc12345",
                          "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                          "rate": "120",
                          "schedule": "month",
                          "status": "active",
                          "nextAction": "pause",
                          "token": "XLM",
                          "createdAt": "2024-01-15T10:00:00.000Z",
                          "updatedAt": "2024-01-16T08:30:00.000Z"
                        }
                      ],
                      "links": { "self": "/api/v1/streams?limit=20" },
                      "meta": { "hasNext": false, "nextCursor": null, "total": 1 }
                    }
                  },
                  "empty-list": {
                    "summary": "No streams exist yet",
                    "value": {
                      "data": [],
                      "links": { "self": "/api/v1/streams?limit=20" },
                      "meta": { "hasNext": false, "nextCursor": null, "total": 0 }
                    }
                  },
                  "paginated-first-page": {
                    "summary": "First page of a multi-page result",
                    "value": {
                      "data": [
                        {
                          "id": "stream-abc12345",
                          "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                          "rate": "50",
                          "schedule": "week",
                          "status": "draft",
                          "nextAction": "start",
                          "token": "XLM",
                          "createdAt": "2024-01-15T10:00:00.000Z",
                          "updatedAt": "2024-01-15T10:00:00.000Z"
                        }
                      ],
                      "links": { "self": "/api/v1/streams?limit=1" },
                      "meta": { "hasNext": true, "nextCursor": "c3RyZWFtLWFiYzEyMzQ1", "total": 3 }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Malformed cursor.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "INVALID_CURSOR",
                    "message": "Malformed cursor",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Too many requests. Retry after 30 seconds.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/500" }
        }
      },
      "post": {
        "operationId": "createStream",
        "summary": "Create a payment stream (v1, deprecated)",
        "description": "Creates a new payment stream in `draft` status. **Deprecated** — use `POST /api/v2/streams` instead.\n\nSupplying an `Idempotency-Key` header makes this request idempotent; replaying the same key and body returns the cached response.",
        "deprecated": true,
        "tags": ["Streams v1"],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Client-generated key (UUID recommended) to make the request idempotent.",
            "schema": { "type": "string", "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["recipient", "rate", "schedule"],
                "properties": {
                  "recipient": {
                    "type": "string",
                    "description": "Stellar public key of the payment recipient (56-character G… address).",
                    "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
                  },
                  "rate": {
                    "type": "string",
                    "description": "Positive decimal payment amount per schedule interval (max 7 decimal places).",
                    "example": "120"
                  },
                  "schedule": {
                    "type": "string",
                    "enum": ["second", "minute", "hour", "day", "week", "month", "year"],
                    "description": "Payment interval.",
                    "example": "month"
                  },
                  "token": {
                    "type": "string",
                    "description": "Stellar asset code (optional, defaults to XLM).",
                    "example": "XLM"
                  }
                }
              },
              "example": {
                "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                "rate": "120",
                "schedule": "month",
                "token": "XLM"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stream created. Initial status is `draft`; call `POST /api/streams/{id}/start` to activate.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "links"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/StreamV1" },
                    "links": {
                      "type": "object",
                      "required": ["self"],
                      "properties": {
                        "self": { "type": "string", "example": "/api/v1/streams/stream-abc12345" }
                      }
                    }
                  }
                },
                "example": {
                  "data": {
                    "id": "stream-abc12345",
                    "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                    "rate": "120",
                    "schedule": "month",
                    "status": "draft",
                    "nextAction": "start",
                    "token": "XLM",
                    "createdAt": "2024-01-15T10:00:00.000Z",
                    "updatedAt": "2024-01-15T10:00:00.000Z"
                  },
                  "links": { "self": "/api/v1/streams/stream-abc12345" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "409": {
            "description": "Idempotency key reused with a different request body.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "IDEMPOTENCY_CONFLICT",
                    "message": "Idempotency key has been used with a different request.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "422": { "$ref": "#/components/responses/422" },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
                "example": {
                  "error": {
                    "code": "RATE_LIMIT_EXCEEDED",
                    "message": "Too many requests. Retry after 30 seconds.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/500" }
        }
      }
    },
...
    "/api/v2/streams": {
      "get": {
        "operationId": "listStreamsV2",
        "summary": "List payment streams (v2)",
        "description": "Returns all streams owned by the authenticated user in the v2 shape (`allowed_actions`, `created_at`, `settlement`).",
        "tags": ["Streams"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Stream list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["streams"],
                  "properties": {
                    "streams": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/StreamV2" }
                    }
                  }
                },
                "examples": {
                  "one-active-stream": {
                    "summary": "Single active stream",
                    "value": {
                      "streams": [
                        {
                          "id": "stream_lf3k9z",
                          "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                          "rate": "120 XLM/month",
                          "status": "active",
                          "allowed_actions": ["pause", "stop"],
                          "created_at": "2024-01-15T10:00:00.000Z",
                          "settlement": null
                        }
                      ]
                    }
                  },
                  "settled-stream": {
                    "summary": "Ended stream with settlement details",
                    "value": {
                      "streams": [
                        {
                          "id": "stream_lf3k9z",
                          "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                          "rate": "120 XLM/month",
                          "status": "ended",
                          "allowed_actions": [],
                          "created_at": "2024-01-15T10:00:00.000Z",
                          "settlement": {
                            "settled_at": "2024-02-01T12:00:00.000Z",
                            "amount": "360 XLM",
                            "currency": "XLM"
                          }
                        }
                      ]
                    }
                  },
                  "empty-list": {
                    "summary": "No streams for this user",
                    "value": { "streams": [] }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401" },
          "500": { "$ref": "#/components/responses/500" }
        }
      },
      "post": {
        "operationId": "createStreamV2",
        "summary": "Create a payment stream (v2)",
        "tags": ["Streams"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["recipient", "rate", "schedule"],
                "properties": {
                  "recipient": {
                    "type": "string",
                    "description": "Stellar public key of the recipient (56-character G… address).",
                    "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
                  },
                  "rate": {
                    "type": "string",
                    "description": "Positive decimal payment amount per schedule interval (max 7 decimal places).",
                    "example": "120"
                  },
                  "schedule": {
                    "type": "string",
                    "enum": ["second", "minute", "hour", "day", "week", "month", "year"],
                    "description": "Payment interval.",
                    "example": "month"
                  },
                  "token": {
                    "type": "string",
                    "description": "Stellar asset code (optional, defaults to XLM).",
                    "example": "XLM"
                  }
                }
              },
              "example": {
                "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                "rate": "120",
                "schedule": "month"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stream created. Initial status is `draft`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StreamV2" },
                "example": {
                  "id": "stream_lf3k9z",
                  "recipient": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG",
                  "rate": "120 XLM/month",
                  "status": "draft",
                  "allowed_actions": ["start"],
                  "created_at": "2024-01-15T10:00:00.000Z",
                  "settlement": null
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "401": { "$ref": "#/components/responses/401" },
          "422": { "$ref": "#/components/responses/422" },
          "500": { "$ref": "#/components/responses/500" }
        }
      }
    },
...
  },
...
}